+++ modules/hanna-react/CHANGELOG.md
@@ -6,6 +6,8 @@
 - `Layout` and `WizardLayout`:
   - feat: Implement `__header__homelink` and `__header__sitename` markup
   - feat: Warn about missing/redundant `siteName` prop
+- fix: Focus error on `SiteSearchAutocomplete` and `AutosuggestSearch`
+  suggestion click
 
 ## 0.10.104